workload_generator: Fix for column parsing for statements with joins
This is a continuation of the effort for the workload generator which was worked on as an intern project.
Previously there was mismatch in the way columns were matched to table in statements with join.
The problems:

Columns were not identified properly in case there were Joins because only 1 table name was considered in a statement.
Columns with tablename.column_name is not handled.
This PR addresses those issues.

Creating a pull with changes for #151053
Fixes: CRDB-51572
epic: none

Release note (cli change):
Join‐table workload columns are now properly qualified.