workload_generator: Fix for column parsing for statements with joins
This is a continuation of the effort for the workload generator which was worked on as an intern project.
Previously there was mismatch in the way columns were matched to table in statements with join.
The problems:

Columns were not identified properly in case there were Joins because only 1 table name was considered in a statement.
Columns with tablename.column_name is not handled.
This PR addresses those issues.

Creating a pull with changes for #151053
Fixes: CRDB-51572
epic: none

Release note (cli change):
Join‐table workload columns are now properly qualified.

workload_generator: skip tables of another database

If a query is run from the current database scope, the queries are marked as the current database even if the query is run for another database. e.g.
SQL> use db1;
SQL> select * from db2.table_in_db2;
The query for db is marked under the database name db1 in the `crdb_internal.node_statement_statistics.txt` file. This causes an issue as unsupported queries also get added for run.

This PR fixes the issue.

Fixes: CRDB-51572
Epic: None

workload_generator: enable workload re-run

Currently, the workload cannot be re-run as the last value of a sequence column is not maintained from the previous run. So, we had to fresh import the data before each run.

In this PR, the last_value is stored in the YAML so that it can be reused in the next run.

There are some additional cleanup changes:
 - Use references instead of pass by value
 - The block for each table was a list, which was unnecessary

Fixes: CRDB-51572
Epic: None