workload_generator: skip tables of another database
If a query is run from the current database scope, the queries are marked as the current database even if the query is run for another database. e.g.
SQL> use db1;
SQL> select * from db2.table_in_db2;
The query for db is marked under the database name db1 in the `crdb_internal.node_statement_statistics.txt` file. This causes an issue as unsupported queries also get added for run.

This PR fixes the issue.

Fixes: CRDB-51572
Epic: None