workload_generator: enable workload re-run
Currently, the workload cannot be re-run as the last value of a sequence column is not maintained from the previous run. So, we had to fresh import the data before each run.

In this PR, the last_value is stored in the YAML so that it can be reused in the next run.

There are some additional cleanup changes:
 - Use references instead of pass by value
 - The block for each table was a list, which was unnecessary

Fixes: CRDB-51572
Epic: None